move into its own namespace
This is useful in case we want to show actual project hover cards in the
future.
