+++ app/components/_index.sass
@@ -20,6 +20,7 @@
 @import "open_project/common/submenu_component"
 @import "filter/filters_component"
 @import "projects/row_component"
+@import "projects/phases/hover_card_component"
 @import "op_primer/border_box_table_component"
 @import "op_primer/form_helpers"
 @import "work_packages/exports/modal_dialog_component"
