+++ app/components/projects/phase_component.rb
@@ -81,7 +81,7 @@ module Projects
 
       {
         hover_card_trigger_target: "trigger",
-        hover_card_url: "/projects/#{phase.id}/hover_card?gate=#{gate}"
+        hover_card_url: hover_card_project_phase_path(phase, gate:)
       }
     end
