+++ app/components/projects/phases/hover_card_component.html.erb
