+++ app/components/projects/phases/hover_card_component.sass
@@ -1,0 +1,3 @@
+.op-hover-card:has(.op-phase-gate-hover-card)
+  max-width: 380px
+
