+++ app/views/project_phases/hover_card/show.html.erb
@@ -1,0 +1,3 @@
+<turbo-frame id="op-hover-card-body">
+  <%= render Projects::Phases::HoverCardComponent.new(id: @id, gate: @gate) %>
+</turbo-frame>
