--- app/views/projects/hover_card/show.html.erb
@@ -1,3 +1,0 @@
-<turbo-frame id="op-hover-card-body">
-  <%= render Projects::HoverCardComponent.new(id: @id, gate: @gate) %>
-</turbo-frame>
